Updated "getidle" to also show who owns the PTY
Knowing who owns the PTY is something I found useful for something else I was playing with. Can probably be cleaned up but works.

Example output:
```
PTY 0 has been idle for 1908696 seconds and is owned by root
PTY 1 has been idle for 321 seconds and is owned by somebody
PTY 2 has been idle for 1 seconds and is owned by root
```

I'm going to try also add some way to have it flag which PTY is *us* because that is useful to know.