add selection of hash or type command in the orc_existsProg
A check for a program or shell command could be executed by a hash or a type command.

The hash command works in the bash and dash but not in the ksh. The type command works in bash, dash and ksh but is not POSIX conform.

The code tries to figure out which command works and defines an alias to the command which could be used in the orc_existsProg function.